Rename Decompiler/Analysis to Annotations and document the metadata product boundary
Acts on the "decompiler + annotations + analysis as a unit" review (B+),
which flagged two seams worth closing:

1. "Analysis" named two unrelated subsystems. The decompiler-side folder
   ILInspector.Decompiler.Analysis (hidden-fact ANNOTATIONS on the IR:
   AnnotationEngine, IHiddenFactClassifier, the Allocation/Unsafety/Lifetime
   classifiers and the annotated/mixed renderers) collided with the separate
   ILInspector.Analysis assembly (whole-assembly memory-safety analysis).
   Rename the folder and namespace to ILInspector.Decompiler.Annotations,
   matching the AnnotationEngine/AnnotationStage vocabulary already in those
   files. Mechanical, compiler-verified rename across the 10 moved files plus
   every consumer (7 test files, IlProjection, MemberCodeProvider, the harness
   AnnotationCheck). Files moved with git mv to preserve history.

2. The three-TypeRef duplication read as drift from the architecture entry
   point. It is in fact a committed product boundary, already argued in full in
   docs/metadata-primitives.md ("Decision (2026-06): stop after step 3" — keep
   ILInspector.Analysis at zero project references, TypeRef unification is
   decisively wrong, tolerate the duplication, with a single trip-wire). The
   gap was discoverability, not the decision. Add a Key Design Decision to
   docs/architecture.md that states the boundary and links to the full
   rationale, and names the Annotations-vs-Analysis distinction so a newcomer
   stops asking "why are there three TypeRefs?".

No behavior change. 350/350 decompiler tests green; markdownlint clean.
