+++ docs/architecture.md
@@ -506,3 +506,5 @@ src/ILInspector.Metadata/   # PE/assembly domain provider
 7. **Automatic TFM selection** — When multiple target frameworks exist, the highest is auto-selected. Override with `--tfm`.
 
 8. **Signature-first output** — Full method signatures with parameter names are the primary output, not just type names, because LLMs need complete information to generate correct code.
+
+9. **Deliberate metadata duplication — `ILInspector.Analysis` is a standalone product** — The whole-assembly memory-safety analyzer (`ILInspector.Analysis`: `LibraryBodyIndex`, `CallTree`, the `Hollow`/`Opaque`/`UnsafeLeverage` classifications) keeps **zero project references** and re-derives its own `TypeRef` / `TypeRefDecoder` / `MemberResolver` rather than sharing the codebase's other type-identity layers (`ILInspector.Metadata`, `ILInspector.MetadataPrimitives`, and the decompiler's `Pipeline/TypeRef`). This is a committed product boundary, not drift: the three `TypeRef` models answer different questions (display string, evidence matching, codegen IR), and `Analysis`'s SRM-direct independence is load-bearing — it ships and evolves as an independent memory-safety deliverable. The full rationale, the rejected consolidation path, and the single trip-wire that would reopen it are in [docs/metadata-primitives.md](metadata-primitives.md) ("Decision (2026-06): stop after step 3"). The two analysis subsystems are also distinct: `ILInspector.Decompiler.Annotations` (hidden-fact *annotations* on the decompiler IR — "where in this method are the allocations/unsafe ops?") versus the `ILInspector.Analysis` assembly ("which methods across this assembly require `unsafe`, and who calls them?").
