+++ src/ILInspector.Decompiler.Tests/AnnotationAnchorTests.cs
@@ -1,5 +1,5 @@
 using System.Collections.Generic;
-using ILInspector.Decompiler.Analysis;
+using ILInspector.Decompiler.Annotations;
 using ILInspector.Decompiler.Pipeline;
 
 namespace ILInspector.Decompiler.Tests;
