+++ src/ILInspector.Decompiler.Tests/AnnotationStructuredViewTests.cs
@@ -1,5 +1,5 @@
 using System.Text.Json;
-using ILInspector.Decompiler.Analysis;
+using ILInspector.Decompiler.Annotations;
 using ILInspector.Decompiler.Pipeline;
 
 namespace ILInspector.Decompiler.Tests;
