+++ src/ILInspector.Decompiler/Annotations/AllocationClassifier.cs
@@ -1,6 +1,6 @@
 using ILInspector.Decompiler.Pipeline;
 
-namespace ILInspector.Decompiler.Analysis;
+namespace ILInspector.Decompiler.Annotations;
 
 /// <summary>
 /// Surfaces heap allocations the C# source hides — boxing, object and array
