+++ src/ILInspector.Decompiler/Annotations/AnnotationAnchor.cs
@@ -1,6 +1,6 @@
 using ILInspector.Decompiler.Pipeline;
 
-namespace ILInspector.Decompiler.Analysis;
+namespace ILInspector.Decompiler.Annotations;
 
 /// <summary>
 /// Anchors offset-keyed annotations onto the statements of a raised tree — the
